map tests from utils to viz tests
